move tests for nodejs to its own package
