+++ docs/Examples.rst
@@ -545,61 +545,61 @@ Typescript
 Typescript Get Data from a Board
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/brainflow_get_data.ts
+.. literalinclude:: ../nodejs_package/tests/brainflow_get_data.ts
    :language: javascript
 
 Typescript Markers
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/markers.ts
+.. literalinclude:: ../nodejs_package/tests/markers.ts
    :language: javascript
 
 Typescript Read Write File
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/serialization.ts
+.. literalinclude:: ../nodejs_package/tests/serialization.ts
    :language: javascript
 
 Typescript Downsample Data
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/downsampling.ts
+.. literalinclude:: ../nodejs_package/tests/downsampling.ts
    :language: javascript
 
 Typescript Transforms
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/transforms.ts
+.. literalinclude:: ../nodejs_package/tests/transforms.ts
    :language: javascript
 
 Typescript Signal Filtering
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/signal_filtering.ts
+.. literalinclude:: ../nodejs_package/tests/signal_filtering.ts
    :language: javascript
 
 Typescript Denoising
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/denoising.ts
+.. literalinclude:: ../nodejs_package/tests/denoising.ts
    :language: javascript
 
 Typescript Band Power
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/bandpower.ts
+.. literalinclude:: ../nodejs_package/tests/bandpower.ts
    :language: javascript
 
 Typescript EEG Metrics
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/eeg_metrics.ts
+.. literalinclude:: ../nodejs_package/tests/eeg_metrics.ts
    :language: javascript
 
 Typescript ICA
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. literalinclude:: ../nodejs_package/brainflow/tests/ica.ts
+.. literalinclude:: ../nodejs_package/tests/ica.ts
    :language: javascript
 
 Notebooks
