+++ nodejs_package/tests/signal_filtering.ts
@@ -1,6 +1,4 @@
-import {BoardShim} from '../board_shim';
-import {BoardIds, FilterTypes} from '../brainflow.types';
-import {DataFilter} from '../data_filter';
+import {BoardIds, BoardShim, DataFilter, FilterTypes} from 'brainflow';
 
 function sleep (ms: number)
 {
