+++ nodejs_package/tsconfig.json
@@ -28,5 +28,5 @@
     "incremental": true
   },
   "include": ["brainflow"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "tests"]
 }
